Move internal/language to language
