+++ cmd/gazelle/BUILD.bazel
@@ -17,13 +17,13 @@ go_library(
     deps = [
         "//config:go_default_library",
         "//internal/flag:go_default_library",
-        "//internal/language:go_default_library",
         "//internal/language/go:go_default_library",
         "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/version:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//language:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
