+++ cmd/gazelle/langs.go
@@ -16,9 +16,9 @@ limitations under the License.
 package main
 
 import (
-	"github.com/bazelbuild/bazel-gazelle/internal/language"
 	"github.com/bazelbuild/bazel-gazelle/internal/language/go"
 	"github.com/bazelbuild/bazel-gazelle/internal/language/proto"
+	"github.com/bazelbuild/bazel-gazelle/language"
 )
 
 var languages = []language.Language{
