+++ internal/language/go/BUILD.bazel
@@ -48,9 +48,9 @@ go_library(
     deps = [
         "//config:go_default_library",
         "//internal/flag:go_default_library",
-        "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
         "//label:go_default_library",
+        "//language:go_default_library",
         "//pathtools:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
@@ -73,12 +73,12 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//config:go_default_library",
-        "//internal/language:go_default_library",
         "//internal/language/proto:go_default_library",
         "//internal/merger:go_default_library",
         "//internal/testtools:go_default_library",
         "//internal/walk:go_default_library",
         "//label:go_default_library",
+        "//language:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
