+++ internal/merger/BUILD.bazel
@@ -17,9 +17,9 @@ go_test(
     srcs = ["merger_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/language:go_default_library",
         "//internal/language/go:go_default_library",
         "//internal/language/proto:go_default_library",
+        "//language:go_default_library",
         "//rule:go_default_library",
     ],
 )
