+++ internal/testtools/config.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/internal/language"
+	"github.com/bazelbuild/bazel-gazelle/language"
 )
 
 // NewTestConfig returns a Config used for tests in any language extension.
